refactor(admin): share the backup test helpers
Three copies of "configure a workspace", "add a CVR with a ballot image", and
"pretend there's plenty of disk space" had already drifted apart, and the
restore tests need the same three.

`addCvrWithBallotImage` now gives each ballot its own batch and CVR file so it
can be called more than once, and returns the image data it wrote so callers
don't restate it.
